pyproject.toml
[project]
name = "pydmi_ritorizo"
version = "0.0.1"
authors = [
  { name="ritorizo" },
]
description = "A library for packing and unpacking byond's dmi files"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent"
]
dependencies = ["pillow>=9.5"]

[project.urls]
"Homepage" = "https://github.com/ritorizo/pydmi"
"Bug Tracker" = "https://github.com/ritorizo/pydmi/issues"

